Progress #1232 -- Packets update (#1231)
*Updated NotifyAddGold, NotifyAddXp, NotifyBlueTip, NotifyChampionDeathTimer, NotifyChampionDie, NotifyChampionRespawn, NotifyLoadScreenPlayerChampion, NotifyLoadScreenPlayerName, NotifyModelUpdate, NotifyPing packets from PacketDefinitions420 to the LeaguePackets Standards.

*Renamed NotifyAnnounceEvent function to NotifyS2C_OnEventWorld (didn't update it yet)

*Updated all references to reflect all the packets that were updated.

*Deleted all unused entries in PacketDefinitions420.

+++ GameServerCore/Packets/PacketDefinitions/Requests/AttentionPingRequest.cs
@@ -6,14 +6,14 @@ namespace GameServerCore.Packets.PacketDefinitions.Requests
     {
         public float X { get; }
         public float Y { get; }
-        public int TargetNetId { get; }
+        public uint TargetNetId { get; }
         public Pings Type { get; }
 
         public AttentionPingRequest(float x, float y, int targetNetId, Pings type)
         {
             X = x;
             Y = y;
-            TargetNetId = targetNetId;
+            TargetNetId = (uint)targetNetId;
             Type = type;
         }
     }
